rafs: move RafsChunkInfo trait to storage module
