+++ src/bin/nydus-image/tree.rs
@@ -28,6 +28,11 @@ use crate::{
     stargz::{self, TocEntry},
 };
 
+// TODO: Import RafsChunkFlags from storage::device is very strange,
+/// perhaps move it to a separate lib someday. And future inspect tool can also
+/// rely on this.
+use rafs::storage::device::RafsChunkFlags;
+
 use nydus_utils::digest::RafsDigest;
 
 use crate::trace::*;
